refactor: use Custom Hook instead of HOC in `Counter.tsx`
